refactor: move to RacingCarPosition in dto package
