+++ src/main/java/domain/RacingCars.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.dto.RacingCarPosition;
+
 import java.util.List;
 import java.util.stream.Collectors;
