+++ src/test/java/RacingCarGameTests/RacingPositionsTests.java
@@ -1,6 +1,6 @@
 package RacingCarGameTests;
 
-import domain.RacingCarPosition;
+import domain.dto.RacingCarPosition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
